+++ CHANGELOG.md
@@ -12,3 +12,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - Issue #3: Triangular grid
+- Issue #5: Resources transformation process